requirements.txt
django>=3.2.0,<4.1.0
djangorestframework>=3.12.0,<3.13.0
pytz
flake8>=4.0.0,<4.1.0
